Correcciones de interfaz e implementacion
Ahora se validan correctamente los costos ingresados para alta premium y
se agregan categorias de botones.
La baja de usuarios ahora comprueba de que pagina viene para retornar
correctamente.
Las reservas ahora muestran todos los casos de fechas posibles.
La ayuda ahora muestra correctamente la lista lateral.
Eliminar couch ahora comprueba que el usuario sea el dueño para evistar
ataques modificando valores en la pagina.
Se agrega funcion para comprobar que no se ingrese valor incorrecto de
monto.
En mis couchs ahora se muestra el estado actual del couch.

+++ couchinn/couchinn/funciones/baja_couch.php
@@ -12,23 +12,31 @@
 	$consulta= "SELECT * FROM couch WHERE Id_Couch='$idcouch' and Visible=1";
 	$consulta_execute = $conexion->query($consulta);
 	if($consulta_execute->num_rows){
-		$consultausuario= "SELECT * FROM usuario WHERE Id_Usuario='$idusuario' and Visible=1";
-		$consulta_execute = $conexion->query($consultausuario);
 		$resultado = $consulta_execute->fetch_assoc();
-		if($password==$resultado['Password']){
-			$actualizacion = "UPDATE couch c left JOIN reserva r ON c.Id_Couch=r.Id_Couch left JOIN comentario o ON c.Id_Couch=o.Id_Couch left JOIN punt_couch p ON c.Id_Couch=p.Id_Couch SET c.Visible = 0, r.Visible=0, o.Visible=0, p.Visible=0  WHERE c.Id_Couch = '$idcouch';";
-			if (mysqli_query($conexion, $actualizacion)) {
-				echo '<script> alert("Se ha eliminado su couch!!!");
-						location.href="../miscouchs.php";
-					</script>';
+		if($resultado['Id_Usuario']==$idusuario){
+			$consultausuario= "SELECT * FROM usuario WHERE Id_Usuario='$idusuario' and Visible=1";
+			$consulta_execute = $conexion->query($consultausuario);
+			$resultado = $consulta_execute->fetch_assoc();
+			if($password==$resultado['Password']){
+				$actualizacion = "UPDATE couch c left JOIN reserva r ON c.Id_Couch=r.Id_Couch left JOIN comentario o ON c.Id_Couch=o.Id_Couch left JOIN punt_couch p ON c.Id_Couch=p.Id_Couch SET c.Visible = 0, r.Visible=0, o.Visible=0, p.Visible=0  WHERE c.Id_Couch = '$idcouch';";
+				if (mysqli_query($conexion, $actualizacion)) {
+					echo '<script> alert("Se ha eliminado su couch!!!");
+							location.href="../miscouchs.php";
+						</script>';
+				}else{
+					echo "ERROR. " . mysqli_error($conexion);
+				}	
 			}else{
-				echo "ERROR. " . mysqli_error($conexion);
-			}	
+				echo '<script> 
+						alert("Verifica tu contraseña actual.");
+						location.href="../eliminarcouch.php?id='.$idcouch.'";
+					</script>';
+			}
 		}else{
-			echo '<script> 
-					alert("Verifica tu contraseña actual.");
-					location.href="../eliminarcouch.php?id='.$idcouch.'";
-				</script>';
+			?>	<script> alert("El couch que intenta borrar no es suyo, metase en sus asuntos.");
+				location.href='../miscouchs.php';
+				</script>
+		<?php
 		}
 	}else{
 		?>	<script> alert("Su couch no existe o ya ha sido eliminado.");
